pyproject.toml
[project]
name = "homunculus"
version = "0.1.0"
description = "Self-hosted personal AI agent for scheduling and coordination"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.52.0",
    "aiohttp>=3.11.0",
    "aiosqlite>=0.21.0",
    "google-api-python-client>=2.160.0",
    "google-auth-oauthlib>=1.2.0",
    "twilio>=9.4.0",
    "structlog>=24.0",
    "opentelemetry-api>=1.30",
    "opentelemetry-sdk>=1.30",
    "opentelemetry-exporter-otlp-proto-http>=1.30",
    "cyclopts>=4.6.0",
    "python-dotenv>=1.2.1",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0",
    "ruff>=0.9.0",
]

[project.scripts]
homunculus = "homunculus.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/homunculus"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
extend-select = [
    "I",       # isort
    "UP",      # pyupgrade
    "B",       # flake8-bugbear
    "SIM",     # flake8-simplify
    "RUF",     # ruff-specific
    "C4",      # flake8-comprehensions
    "PIE",     # flake8-pie
    "PT",      # flake8-pytest-style
    "T20",     # flake8-print
    "PERF",    # perflint
    "PLC0415", # no inline imports
    "TID251",  # banned imports
    "ANN401",  # no typing.Any
    "A",       # flake8-builtins (shadowing builtins)
]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"__future__.annotations".msg = "Not needed on Python 3.14"
